fix(celestia_client): polish env related files
