+++ core/lib/env_config/src/lib.rs
@@ -3,7 +3,7 @@ use serde::de::DeserializeOwned;
 
 mod api;
 mod chain;
-mod clients;
+mod via_celestia;
 mod contract_verifier;
 mod contracts;
 mod database;
